refactor to allow tests and local server to run separately and simultaneously
