+++ functions/server.js
@@ -1,4 +1,4 @@
-const app = require('./index');
+const app = require('./app');
 
 let port = 3300;
